chore : Notice 폴더구조 변경, style 파일 분리, Line 추가
