+++ src/containers/MainPage/components/index.ts
@@ -1,2 +1,3 @@
 export { Search } from './search/Search';
 export { FixedButton } from './fixedButton/FixedButton';
+export { Notice } from './notice/Notice';
